requirements.txt
requests==2.22.0
instabot==0.4.5
python-dotenv==0.10.3
Pillow==6.1.0
